+++ resources/lang/de/app.php
@@ -296,4 +296,6 @@ return [
     'searching' => 'Wird gesucht...',
     'fit_to_page' => 'An Seite anpassen',
     'no_data_registered' => 'Keine Daten registriert.',
+    'not_available' => 'nicht verfügbar',
+    'capacity_exhausted' => 'Kapazität erschöpft',
 ];